// src/core/appSession/types.ts
export interface GroupChatMessage {
  index: number;
  senderMetaId: string;
  timestamp: number;
  content: string;
  encryption?: string;
  protocol?: string;
  pinId?: string;
  groupId?: string;
}